Updates and what to do
Employment has been mostly filled in.

Work through skills first then education, there needs to be a proper layout with special categories before we go through with it

Figure out how to use those buttons in the header to skip to different sections within the page. Ad a contact section within the header to make it take you too the footer, that way users can simply get a hub of contacts