Important Updates
I created a small example of what the landing page could look like, the image is just temporary and more to show what it could look like.

As for the NavBar, it is complete. I am open to suggestions on what would look better.

My next tasks will be to implement some more animations along with the next couple pages/sections.

Next Sections:
- About
- Services